DCES-364: Added @RequestBody annotation to the request parameter of ConcorContributionsService.updateStatus, changed data types of Contribution Id from Long to Integer and status from String to Enum to make the whole flow work as expected.

+++ maat-court-data-api/src/test/java/gov/uk/courtdata/dces/service/ConcorContributionsServiceTest.java
@@ -250,10 +250,10 @@ class ConcorContributionsServiceTest {
     @Test
     void testUpdateConcorContributionsStatus(){
 
-        when(concorRepository.findIdsForUpdate(any())).thenReturn(List.of(1111L,2222L));
+        when(concorRepository.findIdsForUpdate(any())).thenReturn(List.of(1111,2222));
         when(concorRepository.updateStatusForIds(any(), any())).thenReturn(2);
 
-        List<Long> response = concorService.updateConcorContributionStatus(UpdateConcorContributionStatusRequest.builder().recordCount(2)
+        List<Integer> response = concorService.updateConcorContributionStatus(UpdateConcorContributionStatusRequest.builder().recordCount(2)
                 .status(ConcorContributionStatus.SENT).build());
 
         verify(concorRepository).findIdsForUpdate(any());
@@ -266,7 +266,7 @@ class ConcorContributionsServiceTest {
 
         when(concorRepository.findIdsForUpdate(any())).thenReturn(List.of());
 
-        List<Long> response = concorService.updateConcorContributionStatus(UpdateConcorContributionStatusRequest.builder().recordCount(1)
+        List<Integer> response = concorService.updateConcorContributionStatus(UpdateConcorContributionStatusRequest.builder().recordCount(1)
                 .status(ConcorContributionStatus.SENT).build());
 
         verify(concorRepository).findIdsForUpdate(any());
